Consolidate diagnostic builders in src/lcm/regime_building/diagnostics.py.
Follow-up to the #317 split, adjusted for #318's API:
- Keep `_build_compute_intermediates_per_period`, `_wrap_with_reduction`,
  and `_productmap_over_state_action_space` in `diagnostics.py`.
- Extract `get_complete_targets` into `Q_and_F.py` (shared by the
  Q-and-F builder in `processing.py` and the diagnostic builder in
  `diagnostics.py`). No more duplicated helper.
- `processing.py` imports `get_complete_targets` + the builder from
  their proper homes.

Net: processing.py loses ~265 lines; target-enumeration logic is in
one place; cold-path and hot-path builders live in sibling modules.
